+++ src/Oro/Bundle/WebsiteSearchBundle/Attribute/Type/FileSearchableAttributeType.php
@@ -4,12 +4,15 @@ namespace Oro\Bundle\WebsiteSearchBundle\Attribute\Type;
 
 use Oro\Bundle\EntityConfigBundle\Entity\FieldConfigModel;
 
+/**
+ * Attribute type provides metadata for file attribute for search index.
+ */
 class FileSearchableAttributeType extends AbstractSearchableAttributeType
 {
     /**
      * {@inheritdoc}
      */
-    public function getFilterStorageFieldType()
+    public function getFilterStorageFieldTypes(): array
     {
         throw new \RuntimeException('Not supported');
     }
@@ -17,7 +20,7 @@ class FileSearchableAttributeType extends AbstractSearchableAttributeType
     /**
      * {@inheritdoc}
      */
-    public function getSorterStorageFieldType()
+    public function getSorterStorageFieldType(): string
     {
         throw new \RuntimeException('Not supported');
     }
@@ -25,7 +28,7 @@ class FileSearchableAttributeType extends AbstractSearchableAttributeType
     /**
      * {@inheritdoc}
      */
-    public function getFilterType()
+    public function getFilterType(): string
     {
         throw new \RuntimeException('Not supported');
     }
@@ -33,7 +36,7 @@ class FileSearchableAttributeType extends AbstractSearchableAttributeType
     /**
      * {@inheritdoc}
      */
-    public function isLocalizable(FieldConfigModel $attribute)
+    public function isLocalizable(FieldConfigModel $attribute): bool
     {
         return false;
     }
@@ -41,7 +44,7 @@ class FileSearchableAttributeType extends AbstractSearchableAttributeType
     /**
      * {@inheritdoc}
      */
-    public function getFilterableFieldName(FieldConfigModel $attribute)
+    public function getFilterableFieldNames(FieldConfigModel $attribute): array
     {
         throw new \RuntimeException('Not supported');
     }
@@ -49,7 +52,7 @@ class FileSearchableAttributeType extends AbstractSearchableAttributeType
     /**
      * {@inheritdoc}
      */
-    public function getSortableFieldName(FieldConfigModel $attribute)
+    public function getSortableFieldName(FieldConfigModel $attribute): string
     {
         throw new \RuntimeException('Not supported');
     }